Moves about half of the tests from misc and ubpf_vm into ubpf_unified.
