+++ tests/misc.rs
@@ -32,7 +32,7 @@ use solana_rbpf::{
     memory_region::{AccessType, MemoryMapping},
     user_error::UserError,
     verifier::check,
-    vm::{EbpfVm, InstructionMeter, SyscallFunction, SyscallObject},
+    vm::{EbpfVm, InstructionMeter, SyscallObject},
 };
 use std::{fs::File, io::Read, slice::from_raw_parts, str::from_utf8};
 use thiserror::Error;
@@ -105,35 +105,6 @@ use thiserror::Error;
 
 type ExecResult = Result<u64, EbpfError<UserError>>;
 
-macro_rules! test_vm_and_jit {
-    ( $source:tt, $mem:tt, $syscalls:tt, $check:tt ) => {
-        let mut program = assemble($source).unwrap();
-        let syscalls: &[(u32, SyscallFunction<UserError>, usize)] = &$syscalls;
-        for syscall in syscalls {
-            LittleEndian::write_u32(&mut program[syscall.2..syscall.2 + 4], syscall.0);
-        }
-        let executable =
-            EbpfVm::<UserError>::create_executable_from_text_bytes(&program, None).unwrap();
-        let check_closure = $check;
-        {
-            let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &$mem, &[]).unwrap();
-            for syscall in syscalls {
-                vm.register_syscall(syscall.0, syscall.1).unwrap();
-            }
-            assert!(check_closure(vm.execute_program()));
-        }
-        #[cfg(not(windows))]
-        {
-            let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &$mem, &[]).unwrap();
-            for syscall in syscalls {
-                vm.register_syscall(syscall.0, syscall.1).unwrap();
-            }
-            vm.jit_compile().unwrap();
-            assert!(check_closure(unsafe { vm.execute_program_jit() }));
-        }
-    };
-}
-
 fn bpf_syscall_string(
     vm_addr: u64,
     len: u64,
@@ -193,214 +164,6 @@ impl<'a> SyscallObject<UserError> for SyscallWithContext<'a> {
         *self.context = 84;
         Ok(0)
     }
-}
-
-#[test]
-fn test_vm_jit_ldabsb() {
-    test_vm_and_jit!(
-        "
-        ldabsb 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x33 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldabsh() {
-    test_vm_and_jit!(
-        "
-        ldabsh 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x4433 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldabsw() {
-    test_vm_and_jit!(
-        "
-        ldabsw 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x66554433 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldabsdw() {
-    test_vm_and_jit!(
-        "
-        ldabsdw 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0xaa99887766554433 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_err_ldabsb_oob() {
-    test_vm_and_jit!(
-        "
-        ldabsb 0x33
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        {
-            |res: ExecResult| {
-                matches!(res.unwrap_err(),
-                    EbpfError::AccessViolation(pc, access_type, _, _, _)
-                    if access_type == AccessType::Load && pc == 29
-                )
-            }
-        }
-    );
-}
-
-#[test]
-fn test_vm_jit_err_ldabsb_nomem() {
-    test_vm_and_jit!(
-        "
-        ldabsb 0x33
-        exit",
-        [],
-        [],
-        {
-            |res: ExecResult| {
-                matches!(res.unwrap_err(),
-                    EbpfError::AccessViolation(pc, access_type, _, _, _)
-                    if access_type == AccessType::Load && pc == 29
-                )
-            }
-        }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldindb() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x5
-        ldindb r1, 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x88 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldindh() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x5
-        ldindh r1, 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x9988 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldindw() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x4
-        ldindw r1, 0x1
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0x88776655 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_ldinddw() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x2
-        ldinddw r1, 0x3
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        { |res: ExecResult| { res.unwrap() == 0xccbbaa9988776655 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_err_ldindb_oob() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x5
-        ldindb r1, 0x33
-        exit",
-        [
-            0x00, 0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, //
-            0x88, 0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, //
-        ],
-        [],
-        {
-            |res: ExecResult| {
-                matches!(res.unwrap_err(),
-                    EbpfError::AccessViolation(pc, access_type, _, _, _)
-                    if access_type == AccessType::Load && pc == 30
-                )
-            }
-        }
-    );
-}
-
-#[test]
-fn test_vm_jit_err_ldindb_nomem() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x5
-        ldindb r1, 0x33
-        exit",
-        [],
-        [],
-        {
-            |res: ExecResult| {
-                matches!(res.unwrap_err(),
-                    EbpfError::AccessViolation(pc, access_type, _, _, _)
-                    if access_type == AccessType::Load && pc == 30
-                )
-            }
-        }
-    );
 }
 
 /// Error definitions
@@ -685,59 +448,6 @@ fn test_syscall_parameter_on_stack() {
     let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
     vm.register_syscall_ex("log", bpf_syscall_string).unwrap();
     vm.execute_program().unwrap();
-}
-
-#[test]
-fn test_vm_jit_err_syscall_string() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0x0
-        call -0x1
-        mov64 r0, 0x0
-        exit",
-        [72, 101, 108, 108, 111],
-        [(ebpf::hash_symbol_name(b"log"), bpf_syscall_string, 12)],
-        {
-            |res: ExecResult| {
-                matches!(res.unwrap_err(),
-                    EbpfError::AccessViolation(pc, access_type, _, _, _)
-                    if access_type == AccessType::Load && pc == 0
-                )
-            }
-        }
-    );
-}
-
-#[test]
-fn test_vm_jit_syscall_string() {
-    test_vm_and_jit!(
-        "
-        mov64 r2, 0x5
-        call -0x1
-        mov64 r0, 0x0
-        exit",
-        [72, 101, 108, 108, 111],
-        [(ebpf::hash_symbol_name(b"log"), bpf_syscall_string, 12)],
-        { |res: ExecResult| { res.unwrap() == 0 } }
-    );
-}
-
-#[test]
-fn test_vm_jit_syscall() {
-    test_vm_and_jit!(
-        "
-        mov64 r1, 0xAA
-        mov64 r2, 0xBB
-        mov64 r3, 0xCC
-        mov64 r4, 0xDD
-        mov64 r5, 0xEE
-        call -0x1
-        mov64 r0, 0x0
-        exit",
-        [],
-        [(ebpf::hash_symbol_name(b"log"), bpf_syscall_u64, 44)],
-        { |res: ExecResult| { res.unwrap() == 0 } }
-    );
 }
 
 #[test]
@@ -1053,24 +763,4 @@ fn test_fuzz_execute() {
             }
         },
     );
-}
-
-#[cfg(not(windows))]
-#[test]
-#[should_panic(expected = "DivideByZero")]
-fn test_jit_err_div64_by_zero_reg() {
-    let prog = assemble(
-        "
-        mov32 r0, 1
-        mov32 r1, 0
-        div r0, r1
-        exit",
-    )
-    .unwrap();
-    let executable = EbpfVm::<UserError>::create_executable_from_text_bytes(&prog, None).unwrap();
-    let mut vm = EbpfVm::<UserError>::new(executable.as_ref(), &[], &[]).unwrap();
-    vm.jit_compile().unwrap();
-    unsafe {
-        assert_eq!(vm.execute_program_jit().unwrap(), 0);
-    }
 }
